requirements.txt
django==4.2
django-restframework
djangorestframework-simplejwt
Pillow